requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9

qdrant-client>=1.9.0
chromadb>=0.5.0
openai>=1.30.0


numpy>=1.26.0
pandas>=2.2.0
networkx>=3.0

pdfplumber>=0.11.0
python-docx>=1.1.0

reportlab>=4.0.0

tenacity>=8.2.0
loguru>=0.7.0
orjson>=3.10.0
python-dotenv>=1.0.0

sqlalchemy>=2.0.0
alembic>=1.13.0
apscheduler>=3.10,<4.0
filelock>=3.12

python-jose[cryptography]>=3.3.0
httpx>=0.27.0

xfyunsdkspeech>=0.0.4
dashscope>=1.20.0
